Make cosmos node monikers fit in character limit
This allows effectively unlimited subtest structures now.

The fitting algorithm was arbitrarily chosen to keep the start and end
of the test name, and append a hash to the end. It is possible that we
may want a different algorithm at some other point in time, that splits
on / and keeps several segments, so that a reader gets a better sense of
the subtest tree. For now, this approach should suffice.