Make more code aware of *testing.T
In places where we were passing (*testing.T).Name(), prefer to pass the
plain *testing.T value. This makes the interfaces more amenable to
external definitions that are also used specifically in Go tests.

This allows use of t.Cleanup rather than returning callbacks that need
to be deferred.

When we need to sanitize test names for containers, do it later in the
call chains.