WIP: add outline of 'from scratch' test
A root test that shares chains and eventual relayer client state through
subtests will be helpful in onboarding new chains to IBC, and it will
also detect potential flakiness in any early relayer setup.

This is WIP for now because I did not have time to extract helper
functions and a likely helper struct to contain most of this setup and
expose appropriate data to subtests, but this is illustrative of the
general outline on how I would approach this.