Add relayer executions to test report
This was uniform to apply to all the one-off commands, but StartRelayer
and StopRelayer required a slightly different approach due to the
asynchronous style of starting the relayer in the background.

Now, container information, start and stop time, standard output and
standard error, for each relayer execution are included in the report,
so long as the ibc.Relayer implementation has implemented it.