Add test labels
As of this change, labels are associated with individual tests and are
only used for reporting purposes. Filtering tests to run from the
command line is more complex than it might first appear, so that will
happen in a separate change. Label presence in test reports today could
be used for viewing a filtered subset of the report, however.

There are Relayer and Chain labels that are included with top-level
tests, and then individual subtests may have an optional set of labels.