Replace ibctest.SetupTestRun with ibctest.DockerSetup (#91)
* Replace ibctest.SetupTestRun with ibctest.DockerSetup

SetupTestRun was slowing me down every time I needed to read through to
understand what it did.

It returned five values, of which three are not very special: the
context was a plain context.Background, the home directory was the
result of t.TempDir(), and a returned error was intended to always
result in a test failure.

The new ibctest.DockerSetup method returns a *dockertest.Pool and a
networkID value. Because it already accepted a *testing.T, any errors
during Docker setup will cause a call to t.Fatal to immediately fail the
test, as the test cannot continue without those Docker values.

The previously returned context and home directory are instead created
in the calling test where needed. I find this clearer.

This change also inlines the previously exported CreateNetwork function
which is only called internally from DockerSetup.

* Create penumbra home directory earlier in test

The previous ordering and the new ordering both pass on my macOS
workstation but fail on CI. Checking if this order causes CI to pass.