Extract ibctest.chainSet type to simplify future tests
The StartChainsAndRelayerFromFactory function, which has now been
renamed to StartChainPairAndRelayer, had some lengthy sections to set up
chains in preparation for tests. The logic in those sections would apply
just the same to three or more chains. We are working towards supporting
tests beyond a single pair of chains, so extract that common behavior
into the new chainSet type.

The chainSet type is unexported for now, as it contains behavior that
was internal to the StartChainsAndRelayerFromFactory function.