Reorganize test hierarchy
The old way was ROOT/$relayer/$chains/$subtests. The new organization is
ROOT/$arity/$chains/$relayer/$subtests.

In the new layout, $arity currently has one possible value, "pairs".
Although we have discussed the possibility of supporting "triplets" as
an arity, it is not currently on the roadmap for compliance tests.
Leaving arity as part of the test hierarchy keeps that option open
without incurring a command-line API change. (We could also have IBC
configuration tests against a single chain, if that need arises.)

Also of note, this organizes $chains before $relayer. I think this is
more in the spirit of what ibctest is providing. The previous layout was
asking the question, "Does relayer X properly relay for chains A and B,
and for chains C and D?" The new layout asks, "For chains A and B, does
IBC work with relayer X, and for relayer Y?"

To achieve the hierarchy reorganization, the ibctest.ChainFactory
interface was adjusted to replace the Pair method with a Count() method
to inspect how many chains the factory will produce, and a Chains()
method to retrieve all the chains. The previously implemented GetChains
method was removed because a factory will produce a fixed number of
chains; to only accept a subset of the chains would violate the
factory's expectations.