Initial implementation of ibctest.Interchain (#102)
* Add initial ibctest.World type

The World describes a topology of chains and relayers, offering a
declarative interface to test setup.

* Update StartChainPairAndRelayer to use World pattern

This is compiling but failing go test. I am guessing the failure is due
to a hardcoded name in the original set of tests that no longer matches
the dynamic name from the World.

* Fix faucet creation during (*World).Build

Faucets have a fixed name so that preRelayerSetup can be defined before
accounts are created.

This code is still ugly, but tests are passing; chain organization
within the World is coming up next.

* Avoid keying World fields by name

Prefer to key maps by actual ibc.Chain or ibc.Relayer value. The name
abstraction just made everything more complicated.

* Infer chain name from its config

It's unnecessary to provide a second name when one already exists on the
chain's config value.

* Export FaucetAccountKeyName

This seems like it will be necessary for external users.

* Rename World to Interchain

Interchain is a bit more descriptive.

* Move UseDockerNetwork method to ibc.Relayer

The new Interchain type accesses a Relayer instance directly, without
accessing the RelayerFactory. This information is fine for the Relayer
to report directly.

* Extract (*Interchain).genesisWalletAmounts

And drop the InterchainResult type. When callers need to inspect state
modified by the Build method, we will revisit whether it is better to
expose a Result type or to add new introspective methods on the
Interchain struct.

* Extract configureRelayerKeys method