Add more package documentation
My rule of thumb is that a package that already has multiple files will
receive a new doc.go file containing the godoc, even if it is only a
sentence. Otherwise, code authors are left searching through multiple
files when they need to update the package documentation.

If the package only contains a single source file, I add the package
documentation in that lone file.