Add relayer setup conformance test
The test is not yet complete because ibc.Relayer does not yet have a
CreateChannel method -- this was likely an oversight as all other tests
so far simply use CreateLink to create clients, connections, and
channels in one shot.

But this is another example of the Interchain type greatly simplifying
setup for a conformance test.