Add tx_flattened view to blockdb
This view materializes the transactions, joined with blocks, chains, and
test cases.

This is a helpful view for users who are directly interacting with the
sqlite database.

I intend to follow this up with another view that relies on this view,
and parses relevant client/connection/channel details from the tx
objects using sqlite's json functions.