Add FlushPackets and FlushAcknowledgements to ibc.Relayer
The previous ClearQueue method only flushed packets.

Both FlushPackets and FlushAcknowledgements are intended to be run as
one-off commands that do one chunk of work and stop. This is useful for
tests that need to make one unit of progress before making a set of
assertions.

In the future we may add a function like ClearQueue to the test package,
but for now the choices are StartRelayer, or manual flushing.