Add tendermint_events and tendermint_event_attrs tables
The event attributes contain data that is not always present in the
JSON-encoded transactions -- for instance, the packet sequence on an IBC
transfer. The attributes are presented as key-value pairs, and grouped
under a particular type.

This should enable a new view that allows correlating IBC transfers with
their corresponding receives and acknowledgements. It also offers an
alternative to JSON extraction, which could potentially make some other
queries a bit more consistent.