Remove non-empty constraint from attr values
Some events do have an empty value, such as the connection_open_init
type, at least sometimes when the key is counterparty_connection_id.
Disallowing this caused the collector to be stuck in a loop on repeated
failures to save the attribute.

This should fix the consistent failure on the messages view test.