Add TempDir method
This behaves similarly to (*testing.T).TempDir(), except that it
supports runtime configuration to determine whether to remove created
directories for failed tests.

By default, the created directories are removed as part of t.Cleanup. If
the environment variable IBCTEST_KEEP_FAILURE_DIR is set to any
non-empty value, and if the test failed, then the test will log that the
directory is not being cleaned up.

Importers of the ibctest package also have the option to set the package
level variable KeepTempDirOnFailure to opt in to this behavior.