Reduce timeout on starting container for cosmos chain node
This step fails somewhat frequently on my workstation, with 100 tries of
a message like this:

post failed: Post "http://localhost:55493": dial tcp [::1]:55493: connect: connection refused

I'm not sure why it is refusing the connection, but 100 tries with a 5s
sleep between each try is over 8 minutes of waiting to find out that the
test will fail.

Arbitrarily reduce the tries to 40 and the inter-try sleep to 3s, so
that this will only try for about 3 minutes.