Use Docker volumes instead of host mounts for relayer containers (#195)
* Use Docker volumes instead of host mounts for relayer containers

This may help with the hanging Docker containers, or it may not. But it
is importantly a step towards running ibctest against a remote Docker
engine.

This removes the home directory from the Docker Relayer implementations,
as all filesystem operations are now persisted to a short-lived Docker
Volume instead of the host mount.

* Configure docker relayer's volume for relayer's user

The previous approach assumed we were find to simply chmod 0777
everything in the mounted volume, but that failed on Linux due to file
ownership by the root user.

Now, everything should be tailored to the designated relayer user.