Use docker to retrieve genesis hashes
This is one step towards #200, using ephemeral containers to read the
genesis file content from the docker volume instead of reading from the
host disk.

Note that as of this commit, the docker volume is still a host mount,
but it should be seamless to switch this usage to a volume once
everything else for #200 is in place.