penumbra - pclientd integration (#480)
* wip penumbra pclientd

* Updates for 049-pasiphae

* build: bring in ability to pull penumbra protos & gen code

* WIP: implementing missing penumbra methods + smoke test

* WIP: debugging npe in grpc calls

* WIP: balance by addr returns EOF

* update penumbra dep & change rust log level to debug

* WIP: debugging address weirdness

* use raw byte slices in getAddress

we can rather naively convert address strings to byte slices, and not
have to worry about bech32m encoding: the protos will accept the raw
byte slice as sufficient to reconstitute the address.

sprinkled around some debug logging to make sense of what's breaking.

* more debugging around allocations

* fix allocations file format

pd was *silently failing* to process the weirdly constructed csv file.
the golang code was inserting liternal `\n`s rather than newlines.
pd should definitely error on this.

A previous debugging commit touched up the allocations indexing logic,
which wasn't necessary to fix the bug, but sure made things more
comprehensible. This commit restores the original allocations to the
pre-debugging values, aiui.

* handle grpc streaming request for balance

Updates the GetBalanceByAddress request to handle a gRPC streaming
request. Sadly the Penumbra docs aren't very clear that this method
returns a stream; we'll work on that! For now, the
for-loop-and-break-on-EOF seems a sufficiently idiomatic pattern for our
needs. Still to come is making the balance info intelligble.

* fix: get pcli balance checks working + remove debug output

* remove more debug output

* cleanup tx build, auth, broadcast logic + implement func for handling hi/lo amount bytes

* update protos

* filter balances for target denom

* fix proto gen issue in client.pb.go

* get balance checks completely working

* update penumbra 0.57.0, pull new protos, adjust code to account for bigint and rpc changes

* WIP: working on testing local transfers on penumbra

* get balance checks working again + adjust test assertions

* fix get addr and update protos

* update protos and continue work on smoke test

* update to penumbra v0.60.0 and begin cleanup

* more cleanup

* update protos for ictest v8

* update penumbra protos for ibc-go v8 support

* fix: adjust polkadot docker calls to pass env arg

* fix: adjust penumbra implementation with new assumptions in mind

* chore: remove unused code and cleanup comments

* chore: remove unnecessary todos

* chore: use more descriptive homedir for pclientd config files

* chore: implement HomeDir method + utilize const for abci port

---------

Co-authored-by: jtieri <justin@thetieris.com>
Co-authored-by: Conor Schaefer <conor@ruin.dev>
Co-authored-by: Justin Tieri <37750742+jtieri@users.noreply.github.com>