fix: add GRPCConn to PenumbraClientNode and refactor methods
Added a GRPCConn field to PenumbraClientNode to enable persistent gRPC connection. Refactored various functions to use this persistent connection rather than repeatedly dialing a new connection. This streamlines the code, reduces redundancy, and optimizes performance as a new gRPC connection does not need to be opened and closed with every invocation.